Move some parsing code into a cpp file to avoid header include recursion
